+++ tests/PhpPresentation/Tests/Writer/ODPresentation/ContentTest.php
@@ -19,17 +19,17 @@ namespace PhpOffice\PhpPresentation\Tests\Writer\ODPresentation;
 
 use PhpOffice\Common\Drawing as CommonDrawing;
 use PhpOffice\PhpPresentation\PhpPresentation;
+use PhpOffice\PhpPresentation\Shape\Comment;
 use PhpOffice\PhpPresentation\Shape\RichText\Run;
 use PhpOffice\PhpPresentation\Slide\Transition;
 use PhpOffice\PhpPresentation\Style\Alignment;
 use PhpOffice\PhpPresentation\Style\Border;
 use PhpOffice\PhpPresentation\Style\Bullet;
 use PhpOffice\PhpPresentation\Style\Color;
+use PhpOffice\PhpPresentation\Style\Fill;
 use PhpOffice\PhpPresentation\Writer\ODPresentation;
 use PhpOffice\PhpPresentation\Tests\TestHelperDOCX;
 use PhpOffice\Common\Drawing;
-use PhpOffice\PhpPresentation\Style\Fill;
-use PhpOffice\PhpPresentation\Style\PhpOffice\PhpPresentation\Style;
 
 /**
  * Test class for PhpOffice\PhpPresentation\Writer\ODPresentation\Manifest
@@ -39,26 +39,75 @@ use PhpOffice\PhpPresentation\Style\PhpOffice\PhpPresentation\Style;
 class ContentTest extends \PHPUnit_Framework_TestCase
 {
     /**
+     * @var PhpPresentation
+     */
+    protected $oPresentation;
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->oPresentation = new PhpPresentation();
+    }
+
+    /**
      * Executed before each method of the class
      */
     public function tearDown()
     {
+        $this->oPresentation = null;
         TestHelperDOCX::clear();
     }
 
     public function testDrawingWithHyperlink()
     {
-        $phpPresentation = new PhpPresentation();
-        $oSlide = $phpPresentation->getActiveSlide();
+        $oSlide = $this->oPresentation->getActiveSlide();
         $oShape = $oSlide->createDrawingShape();
         $oShape->setPath(PHPPRESENTATION_TESTS_BASE_DIR.'/resources/images/PhpPresentationLogo.png');
         $oShape->getHyperlink()->setUrl('https://github.com/PHPOffice/PHPPresentation/');
     
-        $pres = TestHelperDOCX::getDocument($phpPresentation, 'ODPresentation');
+        $pres = TestHelperDOCX::getDocument($this->oPresentation, 'ODPresentation');
     
         $element = '/office:document-content/office:body/office:presentation/draw:page/draw:frame/office:event-listeners/presentation:event-listener';
         $this->assertTrue($pres->elementExists($element, 'content.xml'));
         $this->assertEquals('https://github.com/PHPOffice/PHPPresentation/', $pres->getElementAttribute($element, 'xlink:href', 'content.xml'));
+    }
+
+    public function testComment()
+    {
+        $expectedName = 'Name';
+        $expectedText = 'Text';
+
+        $oAuthor = new Comment\Author();
+        $oAuthor->setName($expectedName);
+        $oComment = new Comment();
+        $oComment->setAuthor($oAuthor);
+        $oComment->setText($expectedText);
+        $this->oPresentation->getActiveSlide()->addShape($oComment);
+
+        $pres = TestHelperDOCX::getDocument($this->oPresentation, 'ODPresentation');
+        $element = '/office:document-content';
+        $this->assertTrue($pres->elementExists($element, 'content.xml'));
+        $this->assertTrue($pres->attributeElementExists($element, 'xmlns:officeooo', 'content.xml'));
+        $element = '/office:document-content/office:body/office:presentation/draw:page/officeooo:annotation';
+        $this->assertTrue($pres->elementExists($element, 'content.xml'));
+        $element = '/office:document-content/office:body/office:presentation/draw:page/officeooo:annotation/dc:creator';
+        $this->assertTrue($pres->elementExists($element, 'content.xml'));
+        $this->assertEquals($expectedName, $pres->getElement($element, 'content.xml')->nodeValue);
+        $element = '/office:document-content/office:body/office:presentation/draw:page/officeooo:annotation/text:p';
+        $this->assertTrue($pres->elementExists($element, 'content.xml'));
+        $this->assertEquals($expectedText, $pres->getElement($element, 'content.xml')->nodeValue);
+    }
+
+    public function testCommentWithoutAuthor()
+    {
+        $oComment = new Comment();
+        $this->oPresentation->getActiveSlide()->addShape($oComment);
+
+        $pres = TestHelperDOCX::getDocument($this->oPresentation, 'ODPresentation');
+        $element = '/office:document-content/office:body/office:presentation/draw:page/officeooo:annotation';
+        $this->assertTrue($pres->elementExists($element, 'content.xml'));
+        $this->assertFalse($pres->attributeElementExists($element, 'dc:creator', 'content.xml'));
     }
 
     public function testGroup()